fix: use proper method signature and name for on_close method
Previously, the _on_close provided by our module was never called, because the CertStreamClient didn't use the name _on_close. The WebSocketApp used as parent class does provide the method on_close which we can override. This commit changes the name we override from _on_close to on_close. Also we need to change the method signature and some comments and logging parameters.

Despite not being the reason for #4, it still is partly related to it.

fixes #4