fix: use proper method signature for _on_error method
The websocket module changed the method signature some time in the past. The Certstream module has modified the method signature for this method already. When overriding we need to handle that as well, so we need to use three arguments in order to compensate for the self being mandatory, since this is a method of the CertStreamWrapper.

fixes #4